introspection: wait for grace period before dumping kernel state
This change forces atkins to sleep for a fixed amount of time
before dumping the kernel state on test termination. The chosen
amount of sleeping time is assumed to be sufficient for the
kernel to go through a grace period.

By waiting for a grace period, kernel objects whose deletion
are pending and for which the kernel internal references are
all gone during the last grace period will also have been
deleted.

Change-Id: I5d4e54273fa9be9c6035e84108cf4f06ea8d9370