selftest/test_partner - Always set TAP_UUID
In case of single core systems, the cross cpu variants of these tests
are skipped early during SetUp. This however leads to the TAP_UUID never
being set. To solve this we move the corresponding code into a separate
method which can be called in the test after setting the TAP_UUID

Change-Id: Iebe1a448c21941e929de5660f275b48b44e90648