tap/cmdline: fix doxygen parsing problems with val_counter::next()
Doxygen fails to parse the structure variable val_counter. Convert the
code to use a named structure Val_counter, don't use a Doxygen comment
for documenting Val_counter as this private structure is not relevant
for the API and private class members / structures are not part of the
documentation anyway due to `EXTRACT_PRIVATE = NO` in the L4Re Doxygen
configuration.

This avoids the following annoying Doxygen warning:

  ...include/l4/atkins/tap/cmdline:311: warning: no uniquely matching
    class member found for int Atkins::Cmdline::Manager::@265::next()

While at it, also fix the coding style (indentation).

Change-Id: Ib99b99a29101e9b44a7b9c887ec5f1b36ef030d0