atkins/include: Add cov_listener to atkins bundle tests
Atkins bundle tests are another special case of atkins powered test
binaries, where each test defined is executed in a separate thread with
a separate task. For these kinds of tests to dump coverage we need to
attach the Cov_listener, but only in the main task, otherwise (kernel)
coverage would be dumped several times (or just once, but too early).

Change-Id: I52ac484186de32329ce2393f97cde4835eeb9ac7