[API] Add additional infos to Rm-Regions
This shall be useful for debugging facilities.

All regions in a programs address space can be associated with a string,
which is typically a file name and an offset value. This is used by
debugging mechanisms to identify which file is the source of a region.
For that, we extend the L4Re::Rm::attach call with additional optional
argument providing those infos.

The L4Re-VFS is also extended to know a path for a file descriptor.

Storing those infos is optional and a pure debugging mechanism. The
default of this feature is ON.

Change-Id: Ifba9e398016296ccae1d880984f6253e67d847bb